Added an additional tool to help with debugging.
This other tool has 3 function keys
F10 waits for the game to finish loading then tries to start a new game.
F11 while in game it will close out of that (back to main menu) and then wait before trying to start a new game.
F12 is the main feature, it will click what is needed to start a new game from the main menu.

Overall there are bound to be plenty of issues and problems with this script as it was built very quickly.  If you get into trouble use the CTRL+SHIFT+R combo to reset the script.

Also tweaked some settings in RWDebugHelper, not compiled.